+++ Tests/KeyboardKitTests/Mocks/MockInputViewController.swift
@@ -12,11 +12,21 @@ import UIKit
 
 class MockInputViewController: KeyboardInputViewController, Mockable {
     
+    lazy var changeKeyboardTypeRef = MockReference(changeKeyboardType)
     lazy var dismissKeyboardRef = MockReference(dismissKeyboard)
+    lazy var performAutocompleteRef = MockReference(performAutocomplete)
     
     let mock = Mock()
     
+    override func changeKeyboardType(to type: KeyboardType, after delay: DispatchTimeInterval = .milliseconds(0)) {
+        invoke(changeKeyboardTypeRef, args: (type, delay))
+    }
+    
     override func dismissKeyboard() {
         invoke(dismissKeyboardRef, args: ())
+    }
+    
+    override func performAutocomplete() {
+        invoke(performAutocompleteRef, args: ())
     }
 }